+++ src/ts/Test/Unit/EventHandlers/EndScreenEventHandlersTest.ts
@@ -13,15 +13,12 @@ import { Request, INativeResponse } from 'Utilities/Request';
 import { WakeUpManager } from 'Managers/WakeUpManager';
 import { PerformanceAdUnit } from 'AdUnits/PerformanceAdUnit';
 import { Platform } from 'Constants/Platform';
-import { PerformanceCampaign } from 'Models/PerformanceCampaign';
 import { AdUnitContainer, ForceOrientation } from 'AdUnits/Containers/AdUnitContainer';
 import { Activity } from 'AdUnits/Containers/Activity';
 import { ViewController } from 'AdUnits/Containers/ViewController';
 import { StoreName } from "Models/PerformanceCampaign";
 import { Session } from 'Models/Session';
 import { MetaDataManager } from 'Managers/MetaDataManager';
-
-import EndScreenTestPerformanceCampaign1 from 'json/EndScreenTestPerformanceCampaign1.json';
 import { Video } from 'Models/Assets/Video';
 
 describe('EndScreenEventHandlersTest', () => {
@@ -64,9 +61,8 @@ describe('EndScreenEventHandlersTest', () => {
             sinon.stub(sessionManager, 'sendClick').returns(resolvedPromise);
             sinon.spy(nativeBridge.Intent, 'launch');
 
-            const campaignObj = JSON.parse(EndScreenTestPerformanceCampaign1);
             performanceAdUnit = new PerformanceAdUnit(nativeBridge, ForceOrientation.NONE, container, TestFixtures.getPlacement(),
-                new PerformanceCampaign(campaignObj, 'asd', 10), new Video(''), overlay, TestFixtures.getDeviceInfo(Platform.ANDROID), null, endScreen);
+                TestFixtures.getCampaign(), new Video(''), overlay, TestFixtures.getDeviceInfo(Platform.ANDROID), null, endScreen);
         });
 
         it('should send a click with session manager', () => {
@@ -77,9 +73,8 @@ describe('EndScreenEventHandlersTest', () => {
 
         describe('with follow redirects', () => {
             it('with response that contains location, it should launch intent', () => {
-                const campaignObj = JSON.parse(EndScreenTestPerformanceCampaign1);
                 performanceAdUnit = new PerformanceAdUnit(nativeBridge, ForceOrientation.NONE, container, TestFixtures.getPlacement(),
-                    new PerformanceCampaign(campaignObj, 'asd', 10), new Video(''), overlay, TestFixtures.getDeviceInfo(Platform.ANDROID), null, endScreen);
+                    TestFixtures.getCampaignFollowsRedirects(), new Video(''), overlay, TestFixtures.getDeviceInfo(Platform.ANDROID), null, endScreen);
 
                 sinon.stub(sessionManager.getEventManager(), 'clickAttributionEvent').returns(Promise.resolve({
                     url: 'http://foo.url.com',
@@ -99,9 +94,8 @@ describe('EndScreenEventHandlersTest', () => {
             });
 
             it('with response that does not contain location, it should not launch intent', () => {
-                const campaignObj = JSON.parse(EndScreenTestPerformanceCampaign1);
                 performanceAdUnit = new PerformanceAdUnit(nativeBridge, ForceOrientation.NONE, container, TestFixtures.getPlacement(),
-                    new PerformanceCampaign(campaignObj, 'asd', 10), new Video(''), overlay, TestFixtures.getDeviceInfo(Platform.ANDROID), null, endScreen);
+                    TestFixtures.getCampaignFollowsRedirects(), new Video(''), overlay, TestFixtures.getDeviceInfo(Platform.ANDROID), null, endScreen);
 
                 const response = TestFixtures.getOkNativeResponse();
                 response.headers = [];
@@ -133,7 +127,7 @@ describe('EndScreenEventHandlersTest', () => {
             it('should launch market view', () => {
                 sinon.assert.calledWith(<sinon.SinonSpy>nativeBridge.Intent.launch, {
                     'action': 'android.intent.action.VIEW',
-                    'uri': 'market://details?id=fooAppId'
+                    'uri': 'market://details?id=com.iUnity.angryBots'
                 });
             });
 
@@ -172,11 +166,11 @@ describe('EndScreenEventHandlersTest', () => {
             sinon.stub(sessionManager, 'sendClick').returns(resolvedPromise);
             sinon.spy(nativeBridge.UrlScheme, 'open');
 
-            const campaignObj = JSON.parse(EndScreenTestPerformanceCampaign1);
-            campaignObj.store = 'apple';
-            campaignObj.appStoreId = '11111';
+            const campaign = TestFixtures.getCampaign();
+            campaign.set('store', StoreName.APPLE);
+            campaign.set('appStoreId', '11111');
             performanceAdUnit = new PerformanceAdUnit(nativeBridge, ForceOrientation.NONE, container, TestFixtures.getPlacement(),
-                new PerformanceCampaign(campaignObj, 'asd', 10), new Video(''), overlay, TestFixtures.getDeviceInfo(Platform.IOS), null, endScreen);
+                campaign, new Video(''), overlay, TestFixtures.getDeviceInfo(Platform.IOS), null, endScreen);
         });
 
         it('should send a click with session manager', () => {
@@ -189,10 +183,10 @@ describe('EndScreenEventHandlersTest', () => {
         describe('with follow redirects', () => {
             deviceInfo = <DeviceInfo><any>{getOsVersion: () => '9.0'};
             it('with response that contains location, it should open url scheme', () => {
-                const campaignObj = JSON.parse(EndScreenTestPerformanceCampaign1);
-                campaignObj.store = 'apple';
+                const campaign = TestFixtures.getCampaignFollowsRedirects();
+                campaign.set('store', StoreName.APPLE);
                 performanceAdUnit = new PerformanceAdUnit(nativeBridge, ForceOrientation.NONE, container, TestFixtures.getPlacement(),
-                    new PerformanceCampaign(campaignObj, 'asd', 10), new Video(''), overlay, TestFixtures.getDeviceInfo(Platform.IOS), null, endScreen);
+                    campaign, new Video(''), overlay, TestFixtures.getDeviceInfo(Platform.IOS), null, endScreen);
 
                 sinon.stub(sessionManager.getEventManager(), 'clickAttributionEvent').returns(Promise.resolve({
                     url: 'http://foo.url.com',