+++ src/ts/Views/PlayableMRAID.ts
@@ -1,0 +1,302 @@
+import PlayableMRAIDTemplate from 'html/PlayableMRAID.html';
+
+import { NativeBridge } from 'Native/NativeBridge';
+import { Placement } from 'Models/Placement';
+import { MRAIDCampaign } from 'Models/MRAIDCampaign';
+import { Platform } from 'Constants/Platform';
+import { ForceOrientation } from 'AdUnits/Containers/AdUnitContainer';
+import { Template } from 'Utilities/Template';
+import { Localization } from 'Utilities/Localization';
+import { Diagnostics } from 'Utilities/Diagnostics';
+import { MRAIDView } from 'Views/MRAIDView';
+
+export class PlayableMRAID extends MRAIDView {
+
+    private static CloseLength = 30;
+
+    private _localization: Localization;
+
+    private _closeElement: HTMLElement;
+    private _loadingScreen: HTMLElement;
+    private _iframe: HTMLIFrameElement;
+    private _iframeLoaded = false;
+
+    private _messageListener: any;
+    private _loadingScreenTimeout: any;
+    private _prepareTimeout: any;
+
+    private _canClose = false;
+    private _canSkip = false;
+    private _didReward = false;
+
+    private _closeRemaining: number;
+    private _showTimestamp: number;
+
+    constructor(nativeBridge: NativeBridge, placement: Placement, campaign: MRAIDCampaign, language: string) {
+        super(nativeBridge, 'playable-mraid');
+
+        this._placement = placement;
+        this._campaign = campaign;
+        this._localization = new Localization(language, 'endscreen');
+
+        this._template = new Template(PlayableMRAIDTemplate);
+
+        if(campaign) {
+            this._templateData = {
+                'gameName': campaign.getGameName(),
+            };
+            const gameIcon = campaign.getGameIcon();
+            if(gameIcon) {
+                this._templateData.gameIcon = gameIcon.getUrl();
+            }
+            const rating = campaign.getRating();
+            if(rating) {
+                const adjustedRating: number = rating * 20;
+                this._templateData.rating = adjustedRating.toString();
+            }
+            const ratingCount = campaign.getRatingCount();
+            if(ratingCount) {
+                this._templateData.ratingCount = this._localization.abbreviate(ratingCount);
+            }
+        }
+
+        this._bindings = [
+            {
+                event: 'click',
+                listener: (event: Event) => this.onCloseEvent(event),
+                selector: '.close-region'
+            }
+        ];
+    }
+
+    public render() {
+        super.render();
+
+        this._closeElement = <HTMLElement>this._container.querySelector('.close-region');
+        this._loadingScreen = <HTMLElement>this._container.querySelector('.loading-screen');
+
+        const iframe: any = this._iframe = <HTMLIFrameElement>this._container.querySelector('#mraid-iframe');
+
+        this.createMRAID().then(mraid => {
+            iframe.onload = () => this.onIframeLoaded();
+            iframe.srcdoc = mraid;
+        });
+
+        this._messageListener = (event: MessageEvent) => this.onMessage(event);
+        window.addEventListener('message', this._messageListener, false);
+    }
+
+    public show(): void {
+        super.show();
+        this._showTimestamp = Date.now();
+        this.showLoadingScreen();
+    }
+
+    public hide() {
+        this._iframe.contentWindow.postMessage({
+            type: 'viewable',
+            value: false
+        }, '*');
+        if(this._messageListener) {
+            window.removeEventListener('message', this._messageListener, false);
+            this._messageListener = undefined;
+        }
+
+        if(this._loadingScreenTimeout) {
+            clearTimeout(this._loadingScreenTimeout);
+            this._loadingScreenTimeout = undefined;
+        }
+
+        if(this._prepareTimeout) {
+            clearTimeout(this._prepareTimeout);
+            this._prepareTimeout = undefined;
+        }
+        super.hide();
+    }
+
+    public setViewableState(viewable: boolean) {
+        if(this._iframeLoaded && !this._loadingScreenTimeout) {
+            this._iframe.contentWindow.postMessage({
+                type: 'viewable',
+                value: viewable
+            }, '*');
+        }
+    }
+
+    private onIframeLoaded() {
+        this._iframeLoaded = true;
+
+        if(!this._loadingScreenTimeout) {
+            clearTimeout(this._prepareTimeout);
+            this._prepareTimeout = undefined;
+
+            this.showMRAIDAd();
+        }
+    }
+
+    private showLoadingScreen() {
+        this._loadingScreen.style.display = 'block';
+        this._loadingScreenTimeout = setTimeout(() => {
+            if(this._iframeLoaded) {
+                this.showMRAIDAd();
+            } else {
+                // start the prepare timeout and wait for the onload event
+                this._prepareTimeout = setTimeout(() => {
+                    this._canClose = true;
+                    this._closeElement.style.opacity = '1';
+                    this._closeElement.style.display = 'block';
+                    this.updateProgressCircle(this._closeElement, 1);
+
+                    const resourceUrl = this._campaign.getResourceUrl();
+                    Diagnostics.trigger('playable_prepare_timeout', {
+                        'url': resourceUrl ? resourceUrl.getOriginalUrl() : ''
+                    });
+
+                    this._prepareTimeout = undefined;
+                }, 5000);
+            }
+            this._loadingScreenTimeout = undefined;
+        }, 1500);
+    }
+
+    private showMRAIDAd() {
+        if(this._placement.allowSkip()) {
+            const skipLength = this._placement.allowSkipInSeconds();
+            this._closeRemaining = PlayableMRAID.CloseLength;
+            let skipRemaining = skipLength;
+            const updateInterval = setInterval(() => {
+                if(this._closeRemaining > 0) {
+                    this._closeRemaining--;
+                }
+                if(skipRemaining > 0) {
+                    skipRemaining--;
+                    this.updateProgressCircle(this._closeElement, (skipLength - skipRemaining) / skipLength);
+                }
+                if(skipRemaining <= 0) {
+                    this._canSkip = true;
+                    this._closeElement.style.opacity = '1';
+                    this.updateProgressCircle(this._closeElement, 1);
+                }
+                if (this._closeRemaining <= 0) {
+                    clearInterval(updateInterval);
+                    this._canClose = true;
+                }
+            }, 1000);
+        } else {
+            this._closeRemaining = PlayableMRAID.CloseLength;
+            const updateInterval = setInterval(() => {
+                const progress = (PlayableMRAID.CloseLength - this._closeRemaining) / PlayableMRAID.CloseLength;
+                if(progress >= 0.75 && !this._didReward) {
+                    this.onReward.trigger();
+                    this._didReward = true;
+                }
+                if(this._closeRemaining > 0) {
+                    this._closeRemaining--;
+                    this.updateProgressCircle(this._closeElement, progress);
+                }
+                if (this._closeRemaining <= 0) {
+                    clearInterval(updateInterval);
+                    this._canClose = true;
+                    this._closeElement.style.opacity = '1';
+                    this.updateProgressCircle(this._closeElement, 1);
+                }
+            }, 1000);
+        }
+
+        ['webkitTransitionEnd', 'transitionend'].forEach((e) => {
+            if(this._loadingScreen.style.display === 'none') {
+                return;
+            }
+
+            this._loadingScreen.addEventListener(e, () => {
+                this._closeElement.style.display = 'block';
+
+                this._iframe.contentWindow.postMessage({
+                    type: 'viewable',
+                    value: true
+                }, '*');
+
+                this._loadingScreen.style.display = 'none';
+            }, false);
+        });
+
+        this._loadingScreen.classList.add('hidden');
+    }
+
+    private updateProgressCircle(container: HTMLElement, value: number) {
+        const wrapperElement = <HTMLElement>container.querySelector('.progress-wrapper');
+
+        if(this._nativeBridge.getPlatform() === Platform.ANDROID && this._nativeBridge.getApiLevel() < 15) {
+            wrapperElement.style.display = 'none';
+            this._container.style.display = 'none';
+            /* tslint:disable:no-unused-expression */
+            this._container.offsetHeight;
+            /* tslint:enable:no-unused-expression */
+            this._container.style.display = 'block';
+            return;
+        }
+
+        const leftCircleElement = <HTMLElement>container.querySelector('.circle-left');
+        const rightCircleElement = <HTMLElement>container.querySelector('.circle-right');
+
+        const degrees = value * 360;
+        leftCircleElement.style.webkitTransform = 'rotate(' + degrees + 'deg)';
+
+        if(value >= 0.5) {
+            wrapperElement.style.webkitAnimationName = 'close-progress-wrapper';
+            rightCircleElement.style.webkitAnimationName = 'right-spin';
+        }
+    }
+
+    private onCloseEvent(event: Event): void {
+        event.preventDefault();
+        event.stopPropagation();
+        if(this._canSkip && !this._canClose) {
+            this.onSkip.trigger();
+        } else if(this._canClose) {
+            this.onClose.trigger();
+        }
+    }
+
+    private onMessage(event: MessageEvent) {
+        switch(event.data.type) {
+            case 'open':
+                this.onClick.trigger(encodeURI(event.data.url));
+                break;
+            case 'close':
+                this.onClose.trigger();
+                break;
+            case 'orientation':
+                let forceOrientation = ForceOrientation.NONE;
+                switch(event.data.properties.forceOrientation) {
+                    case 'portrait':
+                        forceOrientation = ForceOrientation.PORTRAIT;
+                        break;
+
+                    case 'landscape':
+                        forceOrientation = ForceOrientation.LANDSCAPE;
+                        break;
+
+                    default:
+                        break;
+                }
+                this.onOrientationProperties.trigger({
+                    allowOrientationChange: event.data.properties.allowOrientationChange,
+                    forceOrientation: forceOrientation
+                });
+                break;
+            case 'analyticsEvent':
+                this.onAnalyticsEvent.trigger(event.data.event, (Date.now() - this._showTimestamp) / 1000);
+                break;
+            case 'customMraidState':
+                if(event.data.state === 'completed') {
+                    if(!this._placement.allowSkip() && this._closeRemaining > 5) {
+                        this._closeRemaining = 5;
+                    }
+                }
+                break;
+            default:
+                break;
+        }
+    }
+}